Moved all API calls to one controller
